Split endpoints for reducers
