+++ frontend/src/apis/create-api.ts
@@ -1,34 +1,17 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { UserProfile } from 'models';
+import { ApiTag } from '.';
 
 export const urlString = 'http://127.0.0.1:3002/aurora-instagram-clone/us-central1/api/';
 
-enum ApiTag {
-  PROFILE = 'Profile',
-}
 export interface PostProfile {
   id: string;
   profile: UserProfile;
 }
 
 // Define a service using a base URL and expected endpoints
-export const api = createApi({
+export const emptyApi = createApi({
   baseQuery: fetchBaseQuery({ baseUrl: urlString }),
   tagTypes: [ApiTag.PROFILE],
-  endpoints: builder => ({
-    getProfileById: builder.query<UserProfile, string>({
-      query: id => `/user-profiles/${id}`,
-      providesTags: [ApiTag.PROFILE],
-    }),
-    updateProfileById: builder.mutation<UserProfile, PostProfile>({
-      query: ({ id, profile }) => ({
-        url: `/user-profiles/${id}`,
-        method: 'PATCH',
-        body: profile,
-      }),
-      invalidatesTags: [ApiTag.PROFILE],
-    }),
-  }),
+  endpoints: () => ({}),
 });
-
-export const { useGetProfileByIdQuery, useUpdateProfileByIdMutation } = api;
